Add Admin Get and Purge Events Command
This adds a command line subcommand `eventget` to query for events
and the event matches they triggered. There are some command line
arguments for querying events and limiting the number of results.

The purge subcommand is also added to provide the ability to remove
old events that have been processed.

Signed-off-by: David Brown <dmlb2000@gmail.com>